#402 refresh dashboard on unassessed change, refactor
