+++ src/app/assessments/group/add-assessed-object/add-assessed-object.component.html
@@ -1,0 +1,142 @@
+<div id="assessedObjectsWrapper">
+<hr>
+<md-card>
+    <md-card-title class="pull-left">Assessed Objects</md-card-title>
+    <md-card-title class="pull-right">
+        <button md-fab (click)="addAssessedObject = !addAssessedObject"><i class="material-icons">add</i></button>
+    </md-card-title>
+    <br class="altClearFix">
+</md-card>
+<br>
+
+<div *ngIf="addAssessedObject">
+    <md-card>
+        <md-card-title>Add Assessed Object</md-card-title>
+        <md-card-content>
+            <br>
+            <md-select placeholder="Choose Type" [(ngModel)]="addAssessedType" class="full-width">
+                <md-option value="indicator">Indicator</md-option>
+                <md-option value="course-of-action">Course of Action</md-option>
+                <md-option value="x-unfetter-sensor">Sensor</md-option>
+            </md-select>
+        </md-card-content>
+        <md-card-content *ngIf="addAssessedType">
+            <br>
+            <div [ngSwitch]="addAssessedType">
+
+                <div *ngSwitchCase="'indicator'">
+                    <div class="row">
+                        <div *ngFor="let measurement of indicator.questions" class="col-xs-12">
+                            <label class="mat-label">{{measurement.name | capitalize}}</label>
+                            <md-select [(ngModel)]="measurement.risk" class="full-width mdSelectBotMargin">
+                                <md-option *ngFor="let option of measurement.options" [value]="option.risk">
+                                    {{ option.name }}
+                                </md-option>
+                            </md-select>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <input mdInput placeholder="Name" required [(ngModel)]="indicator.name" value="{{indicator.name}}" />
+                            </md-input-container>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <textarea mdInput placeholder="Description" [(ngModel)]="indicator.description" value="{{indicator.description}}"></textarea>
+                            </md-input-container>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <input mdInput placeholder="Pattern" [(ngModel)]="indicator.pattern" value="{{indicator.pattern}}" />
+                            </md-input-container>
+                        </div>
+                    </div>
+
+                    <div class="row">
+                        <p class="text-right">
+                            <button md-button (click)="addAssessedObject = !addAssessedObject">Cancel</button>
+                            <button md-button [disabled]="!indicator.name" (click)="indicator.name ? createAssessedObject(indicator, currentAttackPattern) : ''">Save</button>
+                        </p>
+                    </div>
+                </div>
+                <!-- End indicator-->
+
+                <div *ngSwitchCase="'course-of-action'">
+                    <div class="row">
+                        <div *ngFor="let measurement of courseOfAction.questions" class="col-xs-12">
+                            <label class="mat-label">{{measurement.name | capitalize}}</label>
+                            <md-select [(ngModel)]="measurement.risk" class="full-width mdSelectBotMargin">
+                                <md-option *ngFor="let option of measurement.options" [value]="option.risk">
+                                    {{ option.name }}
+                                </md-option>
+                            </md-select>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <input mdInput placeholder="Name" required [(ngModel)]="courseOfAction.name" value="{{courseOfAction.name}}">
+                            </md-input-container>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <input mdInput placeholder="Description" [(ngModel)]="courseOfAction.description" value="{{courseOfAction.description}}">
+                            </md-input-container>
+                        </div>
+                    </div>
+
+                    <div class="row">
+                        <p class="text-right">
+                            <button md-button (click)="addAssessedObject = !addAssessedObject">Cancel</button>
+                            <button md-button [disabled]="!courseOfAction.name" (click)="courseOfAction.name ? createAssessedObject(courseOfAction, currentAttackPattern) : ''">Save</button>
+                        </p>
+                    </div>
+                </div>
+                <!-- End COA -->
+
+                <div *ngSwitchCase="'x-unfetter-sensor'">
+                    <div class="row">
+                        <div *ngFor="let measurement of xUnfetterSensor.questions" class="col-xs-12">
+                            <label class="mat-label">{{measurement.name | capitalize}}</label>
+                            <md-select [(ngModel)]="measurement.risk" class="full-width mdSelectBotMargin">
+                                <md-option *ngFor="let option of measurement.options" [value]="option.risk">
+                                    {{ option.name }}
+                                </md-option>
+                            </md-select>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <input mdInput placeholder="Name" required [(ngModel)]="xUnfetterSensor.name" value="{{xUnfetterSensor.name}}" />
+                            </md-input-container>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-12">
+                            <md-input-container class="full-width">
+                                <textarea mdInput placeholder="Description" [(ngModel)]="xUnfetterSensor.description" value="{{xUnfetterSensor.description}}"></textarea>
+                            </md-input-container>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <p class="text-right">
+                            <button md-button (click)="addAssessedObject = !addAssessedObject">Cancel</button>
+                            <button md-button [disabled]="!xUnfetterSensor.name" (click)="xUnfetterSensor.name ? createAssessedObject(xUnfetterSensor, currentAttackPattern) : ''">Save</button>
+                        </p>
+                    </div>
+                </div>
+                <!-- End sensor -->
+
+            </div>
+        </md-card-content>
+    </md-card>
+    <br>
+</div>
