+++ src/app/components/kill-chain-phases/kill-chain-phases.component.ts
